populate_db/embedding_requirements.txt
astrapy==0.7.0
python-dotenv==1.0.0
openai==1.8.0
ftfy==6.2.0
nltk==3.8.1
sentence-transformers==3.0.1
numpy==1.26.4